+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Version 2.0
+- Added a creative drill head that is indestructible, fast, and uses no power to the default drills config
+- Added the ability to create indestructible drills by setting the durability to -1
+
 - Updated the mod to Forge 2847
 - Added a per-drill head energy multiplier
 - Added a (toggleable) drill smart mode to make the drill only use energy when it can actually mine ores